Version 8
* Ajout de la gestion des tarifs
* Finalisation du simulateur de paiement.
* Envois de mail de confirmation (compte, reservation)
* Editeur de PDF
* Ajout de la page de création de compte (admin)
* Inclusion de librarie (PHPMailer, PDFInvoise, FPDF)

+++ views/manage/reservations.php
@@ -12,31 +12,33 @@
 <div class="container">
     <div class="mt-3"></div>
 
-    <table class="table table-striped">
+    <table class="table table-striped text-center">
         <thead>
             <tr>
-                <th>Chambre</th>
+                <th>Chambre N°</th>
+                <th>Etage</th>
+                <th>Capacité</th>
                 <th>Debut</th>
                 <th>Fin</th>
-                <th>Prix</th>
-                <th>Etat</th>
+                <th>Prix TTC</th>
                 <th></th>
             </tr>
         </thead>
         <tbody>
-            <?php foreach( $reservations as $reservation ) {                
-                    $debut = DateTime::createFromFormat('Y-m-d H:i:s', (string)$reservation->debut);
-                    $fin = DateTime::createFromFormat('Y-m-d H:i:s', (string)$reservation->fin);
-                    $days = $fin->diff($debut)->format("%a") + 1;
-            ?>
+            <?php foreach( $reservations as $reservation ) { ?>
                 <tr>
                     <th><?php echo $reservation->numero ?></th>
-                    <th><?php echo $debut->format( 'd-m-y' ) ?></th>
-                    <th><?php echo $fin->format( 'd-m-y' ) ?></th>
-                    <th><?php echo $reservation->prix * $days ?>€</th>
-                    <th><?php echo $reservation->paye == '-1' ? 'Payé' : 'En attente' ?></th>
-                    <th>
-                        <a class="btn btn-danger" href="/manage/deleteReservation/<?php echo $reservation->tid ?>"><i class="fa fa-trash"></i></a>
+                    <th><?php echo $reservation->etage ?></th>
+                    <th><?php echo $reservation->capacite ?></th>
+                    <th><?php echo date( 'd/m/Y', strtotime( $reservation->debut ) ) ?></th>
+                    <th><?php echo date( 'd/m/Y', strtotime( $reservation->fin ) ) ?></th>
+                    <th><?php echo $reservation->prix ?>€</th>
+                    <th class="text-right">
+                        <a class="btn btn-secondary" href="/manage/invoice/<?php echo $reservation->tid ?>" target="_blank"><i class="fa fa-file-pdf"></i></a>
+                        <?php if ( $reservation->paye == 0 ) { ?>
+                            <a class="btn btn-success" href="/billing/buy/<?php echo $reservation->tid ?>">Payer</a>
+                            <a class="btn btn-danger" href="/manage/deleteReservation/<?php echo $reservation->tid ?>"><i class="fa fa-trash"></i></a>
+                        <?php } ?>
                     </th>
                 </tr>
             <?php } ?>
